+++ Content/Items/Reds/ShatteringJustice.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace WellRoundedBalance.Items.Reds
+{
+    internal class ShatteringJustice
+    {
+    }
+}